requirements.txt
# ── core deps ───────────────────────────────────────────
Flask
Flask-SocketIO[eventlet]
eventlet>=0.33           # added explicitly to guarantee presence
paho-mqtt==1.6.1

# ── production server ──────────────────────────────────
gunicorn
